Cleanup and complete generation

+++ src/gen.ts
@@ -1,8 +1,10 @@
 import {writeFileSync} from "node:fs";
+import {readdirSync} from "fs";
 
 function genProblem(n: string): void {
-    const content: string = 'export function run(): number {' +
-        '    return undefined;' +
+    if (readdirSync('challenges').includes(`p${n}.ts`)) throw new Error(`p${n}.ts already exists! Aborting...`);
+    const content: string = 'export function run(): number {\n' +
+        '    return undefined;\n' +
         '};';
     writeFileSync(`challenges/p${n}.ts`, content, {encoding: 'utf8'});
 }
